Max Updates v4.11 - You can now set the grid spacing size to follow the single object's width or height using the resize buttons in the Tilemap Panel respectively (individually or you can lock both to the same values), OR you can scale/resize the single selected object to match the grid size's width, height, or both, in the right click menu of the object (only activated when tilemap is on and single object is selected).

+++ ZodiaClash/HighOctane_CustomEngine/Global.cpp
@@ -86,10 +86,10 @@ Entity keyObjectID{ std::numeric_limits<Entity>().max() };
 glm::vec4 keyObjectColor{ RESET_VEC4 };
 bool snappingOn{ true };
 bool tilemapOn{ false };
-extern int gridSpacingX{ 200 };
-extern int gridSpacingY{ 200 };
-extern int gridOffsetX{ 0 };
-extern int gridOffsetY{ 0 };
+extern float gridSpacingX{ 200.f };
+extern float gridSpacingY{ 200.f };
+extern float gridOffsetX{ 0.f };
+extern float gridOffsetY{ 0.f };
 extern int gridAlignment{ SA_CENTER };
